requirements.txt
websockets>=12.0
requests>=2.31.0
aiohttp>=3.9.0
pandas>=2.0.0
numpy>=1.24.4
python-dotenv>=1.0.0
tenacity>=8.2.0
pytz>=2023.3
rich>=13.6.0
pydantic>=2.6.0
py-clob-client==0.28.0
python-dateutil>=2.9.0
